refactor(shm): extract ChannelLayout to layout.rs (business-side, not core)
Step 2 of SHM business-decoupling. `ChannelLayout` and `allocate_layouts`
encode the (channel_id, point_type, point_id) → slot translation. They
know the 4 PointType semantics (T/S/C/A), which is business. They have
no place in `unified_shm.rs` alongside the on-mmap `UnifiedHeader` /
Reader / Writer mechanism.

Move them to a dedicated `layout.rs` so the business / infra split is
visible at the file level: `unified_shm.rs` now only owns the mmap
header layout and pure read/write/snapshot mechanics; `layout.rs` owns
the channel→slot index translation.

Public re-exports preserved (top-level `ChannelLayout` /
`allocate_layouts` still available); zero external API changes.

Cross-process layout agreement remains via the existing `routing_hash`
field in `UnifiedHeader` — that field is already a manifest-style
fingerprint, no header layout change is needed here.
